+++ Models/Dosificador/DatoModel.cs
@@ -9,6 +9,6 @@ namespace Models.Dosificador
     public class DatoModel
     {
         public int CantidadNotasRemision { get; set; }
-        public int SumaCantidadesNotaRemision { get; set; }
+        public decimal SumaCantidadesNotaRemision { get; set; }
     }
 }
